requirements_vps.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
SQLAlchemy==2.0.35
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.3
email-validator==2.2.0
jinja2==3.1.4
python-multipart==0.0.9
